Move source and make build with latest Skia (#118)
Patch by @sunflowerdeath, refined to build with C++11 as well on none Skia based ports.
* move all public headers into `include/svgnative`
* make all headers in `src` private
* update to use skia m85
* allow to change skia source location using var `SKIA_SOURCE_DIR`
* Enable C++14 for Skia rendering port only
* Show C++ standard in CMake output
* std::tuple has no constexpr ctor, use struct for CSSColorInfo again
