+++ svgnative/example/testGDIPlus/TestGDIPlus.cpp
@@ -18,8 +18,8 @@ governing permissions and limitations under the License.
 #include <Gdiplus.h>
 #include <memory>
 
-#include "SVGDocument.h"
-#include "GDIPlusSVGRenderer.h"
+#include "svgnative/SVGDocument.h"
+#include "svgnative/ports/gdiplus/GDIPlusSVGRenderer.h"
 
 /* We need to include the Gdiplus lib */
 #pragma comment (lib, "Gdiplus.lib")
