+++ svgnative/include/svgnative/SVGDocument.h
