+++ svgnative/include/svgnative/SVGRenderer.h
