+++ svgnative/include/svgnative/ports/cairo/CairoSVGRenderer.h
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 Adobe. All rights reserved.
+Copyright 2019 suzuki toshiya <mpsuzuki@hiroshima-u.ac.jp>. All rights reserved.
 This file is licensed to you under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License. You may obtain a copy
 of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -10,19 +10,25 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-#ifndef SVGViewer_GDIPlusSVGRenderer_h
-#define SVGViewer_GDIPlusSVGRenderer_h
+#ifndef SVGViewer_CairoSVGRenderer_h
+#define SVGViewer_CairoSVGRenderer_h
 
-#include "SVGRenderer.h"
-#include <gdiplus.h>
+#include <list>
+#include "svgnative/SVGRenderer.h"
+#include "cairo.h"
 
 namespace SVGNative
 {
-class GDIPlusSVGPath final : public Path
+
+  // SkiaSVGPath object is able to be amended, but Cairo has no API to append something
+  // to existing cairo_path_t object. Thus, we hold cairo_t object. Its surface could
+  // be retrieved in later, by cairo_get_target().
+
+class CairoSVGPath final : public Path
 {
 public:
-    GDIPlusSVGPath();
-    ~GDIPlusSVGPath();
+    CairoSVGPath();
+    ~CairoSVGPath();
 
     void Rect(float x, float y, float width, float height) override;
     void RoundedRect(float x, float y, float width, float height, float rx, float ry) override;
@@ -34,18 +40,17 @@ public:
     void CurveToV(float x2, float y2, float x3, float y3) override;
     void ClosePath() override;
 
-    const Gdiplus::GraphicsPath& GetGraphicsPath() const;
+    cairo_t* mPathCtx{};
 
 private:
-    Gdiplus::GraphicsPath mPath;
     float mCurrentX{};
     float mCurrentY{};
 };
 
-class GDIPlusSVGTransform final : public Transform
+class CairoSVGTransform final : public Transform
 {
 public:
-    GDIPlusSVGTransform(float a, float b, float c, float d, float tx, float ty);
+    CairoSVGTransform(float a, float b, float c, float d, float tx, float ty);
 
     void Set(float a, float b, float c, float d, float tx, float ty) override;
     void Rotate(float r) override;
@@ -53,42 +58,38 @@ public:
     void Scale(float sx, float sy) override;
     void Concat(float a, float b, float c, float d, float tx, float ty) override;
 
-    const Gdiplus::Matrix& GetMatrix() const;
-
-private:
-    Gdiplus::Matrix mTransform;
+    cairo_matrix_t mMatrix;
 };
 
-class GDIPlusSVGImageData final : public ImageData
+class CairoSVGImageData final : public ImageData
 {
 public:
-    GDIPlusSVGImageData(const std::string& base64, ImageEncoding encoding);
-    ~GDIPlusSVGImageData();
+    CairoSVGImageData(const std::string& base64, ImageEncoding encoding);
+    ~CairoSVGImageData();
 
     float Width() const override;
+
     float Height() const override;
 
-    const std::unique_ptr<Gdiplus::Image>& GetImage() const;
 
-private:
-    std::unique_ptr<Gdiplus::Image> mImage;
+    cairo_surface_t* mImageData{};
 };
 
-class SVG_IMP_EXP GDIPlusSVGRenderer final : public SVGRenderer
+class SVG_IMP_EXP CairoSVGRenderer final : public SVGRenderer
 {
 public:
-    GDIPlusSVGRenderer();
+    CairoSVGRenderer();
+    ~CairoSVGRenderer();
 
-    virtual ~GDIPlusSVGRenderer() 
-    { 
-    }
+    std::unique_ptr<ImageData> CreateImageData(const std::string& base64, ImageEncoding encoding) override { return std::unique_ptr<CairoSVGImageData>(new CairoSVGImageData(base64, encoding)); }
 
-    std::unique_ptr<ImageData> CreateImageData(const std::string& base64, ImageEncoding encoding) override;
-
-    std::unique_ptr<Path> CreatePath() override;
+    std::unique_ptr<Path> CreatePath() override { return std::unique_ptr<CairoSVGPath>(new CairoSVGPath); }
 
     std::unique_ptr<Transform> CreateTransform(
-        float a = 1.0, float b = 0.0, float c = 0.0, float d = 1.0, float tx = 0.0, float ty = 0.0) override;
+        float a = 1.0, float b = 0.0, float c = 0.0, float d = 1.0, float tx = 0.0, float ty = 0.0) override
+    {
+        return std::unique_ptr<CairoSVGTransform>(new CairoSVGTransform(a, b, c, d, tx, ty));
+    }
 
     void Save(const GraphicStyle& graphicStyle) override;
     void Restore() override;
@@ -96,23 +97,12 @@ public:
     void DrawPath(const Path& path, const GraphicStyle& graphicStyle, const FillStyle& fillStyle, const StrokeStyle& strokeStyle) override;
     void DrawImage(const ImageData& image, const GraphicStyle& graphicStyle, const Rect& clipArea, const Rect& fillArea) override;
 
-    void SetGraphicsContext(Gdiplus::Graphics* inContext)
-    {
-        mContext = inContext;
-    }
-
-    void ReleaseGraphicsContext()
-    {
-        mContext = nullptr;
-    }
+    void SetCairo(cairo_t* cairo);
 
 private:
-    std::unique_ptr<Gdiplus::Brush> CreateGradientBrush(const Gradient& gradient, float opacity);
-
-    Gdiplus::Graphics* mContext{};
-    std::vector<Gdiplus::GraphicsState> mStateStack;
+    cairo_t* mCairo{};
 };
 
 } // namespace SVGNative
 
-#endif // SVGViewer_GDIPlusSVGRenderer_h
+#endif // SVGViewer_CairoSVGRenderer_h
