+++ svgnative/src/DeprecatedSVGDocumentImpl.cpp
@@ -12,7 +12,7 @@ governing permissions and limitations under the License.
 
 #ifdef STYLE_SUPPORT
 #include "SVGDocumentImpl.h"
-#include "SVGDocument.h"
+#include "svgnative/SVGDocument.h"
 #include "xml/XMLParser.h"
 
 using namespace SVGNative::xml;
