+++ svgnative/src/SVGNativeCWrapper.cpp
@@ -10,29 +10,29 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-#include "SVGNativeCWrapper.h"
-#include "SVGDocument.h"
-#include "SVGRenderer.h"
+#include "svgnative/SVGNativeCWrapper.h"
+#include "svgnative/SVGDocument.h"
+#include "svgnative/SVGRenderer.h"
 #ifdef USE_CAIRO
-#include "CairoSVGRenderer.h"
+#include "svgnative/ports/cairo/CairoSVGRenderer.h"
 #endif
 #ifdef USE_CG
-#include "CGSVGRenderer.h"
+#include "svgnative/ports/cg/CGSVGRenderer.h"
 #include <ApplicationServices/ApplicationServices.h>
 #include <CoreGraphics/CoreGraphics.h>
 #endif
 #ifdef USE_SKIA
-#include "SkiaSVGRenderer.h"
+#include "svgnative/ports/skia/SkiaSVGRenderer.h"
 #include "SkCanvas.h"
 #include "SkImage.h"
 #endif
 #ifdef USE_GDIPLUS
 #include <windows.h>
 #include <Gdiplus.h>
-#include "GDIPlusSVGRenderer.h"
+#include "svgnative/ports/gdiplus/GDIPlusSVGRenderer.h"
 #endif
 #ifdef USE_TEXT
-#include "StringSVGRenderer.h"
+#include "svgnative/ports/string/StringSVGRenderer.h"
 #endif
 
 #include <map>
