+++ svgnative/src/ports/cairo/CairoImageInfo.c
