+++ svgnative/src/ports/cairo/CairoImageInfo.h
